writing out patched [Tag<Data>] no longer uses GitTagData
it's just an extension on a `URL` now.
